Delete from custom list

+++ app.js
@@ -2,6 +2,7 @@ const express = require("express");
 const bodyParser = require("body-parser");
 // const date = require(__dirname + "/date.js");
 const mongoose = require("mongoose");
+const _ = require("lodash");
 
 const app = express();
 
@@ -95,7 +96,7 @@ app.get("/", function (req, res) {
 });
 
 app.get("/:customListName", function(req, res){
-  const customListName = req.params.customListName;
+  const customListName = _.capitalize(req.params.customListName);
 
   List.findOne({name: customListName})
     .then(function(foundList){
@@ -148,16 +149,22 @@ app.post("/", function (req, res) {
 
 app.post("/delete", function(req, res){
   const checkedItemId = req.body.checkbox;
+  const listName = req.body.listName;
 
-  Item.findByIdAndRemove(checkedItemId)
-    .then(function(foundItem){
-      Item.deleteOne({_id: checkedItemId});
-      console.log("Successfully deleted checked item.");
-      res.redirect("/");
-    })
-    .catch(function(err){
-      console.log(err)
-    })
+  if(listName === "Today"){
+    Item.findByIdAndRemove(checkedItemId)
+      .then(function(foundItem){
+        Item.deleteOne({_id: checkedItemId});
+        console.log("Successfully deleted checked item.");
+        res.redirect("/");
+      })
+  }
+  else {
+    List.findOneAndUpdate({name: listName}, {$pull: {items: {_id: checkedItemId}}})
+      .then(function(foundList){
+        res.redirect(`/${listName}`)
+      });
+  };
 });
 
 app.get("/work", function (req, res) {
